Add Frontend product options
