+++ package.json
@@ -30,6 +30,7 @@
     "grunt-checktextdomain": "^1.0.1",
     "grunt-wp-i18n": "^1.0.3",
     "grunt-wp-readme-to-markdown": "^2.1.0",
+    "load-grunt-tasks": "^5.1.0",
     "webpack-remove-empty-scripts": "^1.0.4"
   },
   "stylelint": {
